NXP-29738: Fix S3 direct upload same file import

+++ modules/core/nuxeo-core-binarymanager-cloud/nuxeo-core-binarymanager-s3/src/main/java/org/nuxeo/ecm/core/storage/sql/S3DirectBatchHandler.java
@@ -63,6 +63,7 @@ import com.amazonaws.client.builder.AwsClientBuilder;
 import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
+import com.amazonaws.services.s3.model.AmazonS3Exception;
 import com.amazonaws.services.s3.model.CopyObjectRequest;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.SSEAwsKeyManagementParams;
@@ -254,11 +255,12 @@ public class S3DirectBatchHandler extends AbstractBatchHandler {
         ObjectMetadata metadata = amazonS3.getObjectMetadata(bucket, fileKey);
 
         String key;
+        String eTag = metadata.getETag();
 
         if (!useDeDuplication()) {
             key = StringUtils.removeStart(fileKey, bucketPrefix);
         } else {
-            key = metadata.getETag();
+            key = eTag;
             if (isEmpty(key)) {
                 return false;
             }
@@ -279,35 +281,28 @@ public class S3DirectBatchHandler extends AbstractBatchHandler {
                 // TODO SSE-C
             }
 
-            Copy copy = getTransferManager().copy(copyObjectRequest);
-            try {
-                copy.waitForCompletion();
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new NuxeoException(e);
-            } finally {
-                amazonS3.deleteObject(bucket, fileKey);
-            }
-
-            metadata = amazonS3.getObjectMetadata(bucket, bucketKey);
+            eTag = doMove(copyObjectRequest);
 
             // if we did a multipart upload but can do a non multipart copy we can get back the digest as key
             boolean isMultipartUpload = key.matches(".+-\\d+$");
             boolean canDoNonMultipartCopy = metadata.getContentLength() < getTransferManager().getConfiguration()
                                                                                               .getMultipartCopyThreshold();
             if (isMultipartUpload && canDoNonMultipartCopy) {
-                key = metadata.getETag();
+                key = eTag;
                 String previousBucketKey = bucketKey;
                 bucketKey = bucketPrefix + key;
-                CopyObjectRequest renameRequest = new CopyObjectRequest(bucket, previousBucketKey, bucket, bucketKey);
-                Copy rename = getTransferManager().copy(renameRequest);
-                try {
-                    rename.waitForCompletion();
-                } catch (InterruptedException e) {
-                    Thread.currentThread().interrupt();
-                    throw new NuxeoException(e);
-                } finally {
-                    amazonS3.deleteObject(bucket, previousBucketKey);
+                if (amazonS3.doesObjectExist(bucket, bucketKey)) {
+                    // another thread has uploaded the same blob and has done the move
+                    // clean up remaining S3 object if needed
+                    try {
+                        amazonS3.deleteObject(bucket, previousBucketKey);
+                    } catch (AmazonS3Exception e) {
+                        log.debug("Unable to cleanup object, move has already been done", e);
+                    }
+                } else {
+                    // move S3 object to eTag as key
+                    var renameRequest = new CopyObjectRequest(bucket, previousBucketKey, bucket, bucketKey);
+                    eTag = doMove(renameRequest);
                 }
             }
         }
@@ -318,7 +313,7 @@ public class S3DirectBatchHandler extends AbstractBatchHandler {
         blobInfo.filename = fileInfo.getFilename();
         blobInfo.length = metadata.getContentLength();
         blobInfo.key = key;
-        blobInfo.digest = defaultString(metadata.getContentMD5(), metadata.getETag());
+        blobInfo.digest = defaultString(metadata.getContentMD5(), eTag);
 
         Blob blob;
 
@@ -337,6 +332,19 @@ public class S3DirectBatchHandler extends AbstractBatchHandler {
         }
 
         return true;
+    }
+
+    protected String doMove(CopyObjectRequest request) {
+        Copy rename = getTransferManager().copy(request);
+        try {
+            var copyResult = rename.waitForCopyResult();
+            return copyResult.getETag();
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+            throw new NuxeoException(e);
+        } finally {
+            amazonS3.deleteObject(bucket, request.getSourceKey());
+        }
     }
 
     protected BlobProvider getBlobProvider() {
